New UndoHistory entity
