+++ app/javascript/Domain/UseCases/Players/AdjustCash.js
@@ -1,21 +1,10 @@
 export default class AdjustCash {
-  #game;
+  static changeBy(player, amount, undoHistory) {
+    const originalAmount = player.cash;
+    undoHistory?.pushUndoOperation(() => { player.cash = originalAmount; });
 
-  constructor(game) {
-    this.#game = game;
-  }
-
-  static changeBy(player, amount) {
     if (amount) {
       player.cash = Math.max(0, player.cash + amount);
-    }
-  }
-  undoableChangeBy(player, amount) {
-    if (amount) {
-      const originalAmount = player.cash;
-      this.#game.pushUndoOperation(() => { player.cash = originalAmount; });
-
-      AdjustCash.changeBy(player, amount);
     }
   }
 }
